Move table inside of Simulation
